Implemented partitions for web master pull job.
