+++ gobblin-core/src/main/java/gobblin/source/extractor/extract/google/webmaster/GoogleWebMasterSource.java
@@ -12,7 +12,6 @@ public class GoogleWebMasterSource extends QueryBasedSource<String, String[]> {
   public static final String KEY_PROPERTY = "source.google_webmasters.property";
   public static final String KEY_PREDEFINED_FILTERS = "source.google_webmasters.predefined_filters";
   public static final String KEY_API_SCOPE = "source.google.api.scope";
-  public static final String KEY_REQUEST_DATE = "source.google_webmasters.request.date";
   public static final String KEY_COUNTRIES = "source.google_webmasters.request.countries";
   public static final String KEY_REQUEST_PAGE_LIMIT = "source.google_webmasters.request.page_limit";
   public static final String KEY_REQUEST_QUERY_LIMIT = "source.google_webmasters.request.query_limit";
